backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
sqlalchemy==2.0.21
pydantic==1.10.12
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
jose==3.3.0
